Wrap MultiSmartContractVerifier into CachedContractSignatureVerifier
- Create a CacheContractSifnatureVerifier which implements SmartContractSignatureVerifier
- Move all the logic to mls_validation_service
